Support glyph indices beyond 65535 throughout font subsystem
Extended font subsystem from 16-bit to 32-bit glyph indices to support
large fonts (massive CJK fonts, emoji fonts with many variants).

Core Changes:
- Changed FontFile.CharacterToGlyphIndex from Dictionary<int, ushort> to Dictionary<int, uint>
- Updated GetGlyphIndex() return type from ushort? to uint?
- Updated ShapedGlyph.GlyphId from ushort to uint

Implementation Details:
1. **FontFile.cs**: Updated all glyph index handling to use uint
   - GetAdvanceWidth, GetKerning, GetGlyphData methods updated
   - Added bounds checking for legacy kerning (still uses ushort pairs)
   - Documented that kern table doesn't support glyphs > 65535

2. **CmapTableParser.cs**: Removed 0xFFFF limitation in ParseFormat12
   - Now stores full range of glyph indices from format 12 tables
   - Removed TODO about supporting glyphs beyond 65535

3. **FontSubsetter.cs**: Updated glyph mapping
   - Changed BuildGlyphMapping to use Dictionary<uint, ushort>
   - Input font can have uint indices, output subset uses ushort (always < 65536 glyphs)
   - Updated all iteration and array indexing with proper casts

4. **OpenTypeShaper.cs**: Updated text shaping
   - Changed ShapedGlyph.GlyphId to uint
   - Added bounds checking for GSUB/GPOS operations (use 16-bit indices)
   - Glyphs > 65535 skip ligatures, substitutions, and positioning
   - This is correct: OpenType GSUB/GPOS tables use 16-bit glyph IDs

5. **TrueTypeFontSerializer.cs**: Added explicit casts
   - Format 4 cmap tables use 16-bit indices
   - Safe for subset fonts which always have < 65536 glyphs

Compatibility Notes:
- Legacy tables (kern, GSUB, GPOS) still use 16-bit glyph indices per spec
- Fonts with > 65535 glyphs won't have kerning/ligatures for high-index glyphs
- This is acceptable: such massive fonts are extremely rare

Files modified:
- src/Folly.Fonts/Models/FontFile.cs
- src/Folly.Fonts/Tables/CmapTableParser.cs
- src/Folly.Fonts/FontSubsetter.cs
- src/Folly.Fonts/OpenType/OpenTypeShaper.cs
- src/Folly.Fonts/TrueTypeFontSerializer.cs

+++ src/Folly.Fonts/FontSubsetter.cs
@@ -42,10 +42,11 @@ public class FontSubsetter
     /// <summary>
     /// Builds a mapping from old glyph indices to new glyph indices in the subset.
     /// Always includes glyph 0 (.notdef) as required by the TrueType spec.
+    /// Supports fonts with glyph indices beyond 65535.
     /// </summary>
-    private static Dictionary<ushort, ushort> BuildGlyphMapping(FontFile font, HashSet<char> usedCharacters)
+    private static Dictionary<uint, ushort> BuildGlyphMapping(FontFile font, HashSet<char> usedCharacters)
     {
-        var glyphMapping = new Dictionary<ushort, ushort>();
+        var glyphMapping = new Dictionary<uint, ushort>();
         ushort newGlyphIndex = 0;
 
         // Always include glyph 0 (.notdef) - required by TrueType spec
@@ -68,7 +69,7 @@ public class FontSubsetter
     /// <summary>
     /// Creates a new FontFile containing only the subset of glyphs.
     /// </summary>
-    private static FontFile CreateSubsetFontFile(FontFile originalFont, Dictionary<ushort, ushort> glyphMapping)
+    private static FontFile CreateSubsetFontFile(FontFile originalFont, Dictionary<uint, ushort> glyphMapping)
     {
         var subsetFont = new FontFile
         {
@@ -97,7 +98,7 @@ public class FontSubsetter
         };
 
         // Build reverse mapping for character lookups
-        var newCharToGlyphIndex = new Dictionary<int, ushort>();
+        var newCharToGlyphIndex = new Dictionary<int, uint>();
         foreach (var kvp in originalFont.CharacterToGlyphIndex)
         {
             if (glyphMapping.TryGetValue(kvp.Value, out ushort newIndex))
@@ -113,15 +114,15 @@ public class FontSubsetter
 
         foreach (var kvp in glyphMapping)
         {
-            ushort oldIndex = kvp.Key;
+            uint oldIndex = kvp.Key;
             ushort newIndex = kvp.Value;
 
-            if (oldIndex < originalFont.GlyphAdvanceWidths.Length)
+            if (oldIndex < (uint)originalFont.GlyphAdvanceWidths.Length)
             {
                 subsetFont.GlyphAdvanceWidths[newIndex] = originalFont.GlyphAdvanceWidths[oldIndex];
             }
 
-            if (oldIndex < originalFont.GlyphLeftSideBearings.Length)
+            if (oldIndex < (uint)originalFont.GlyphLeftSideBearings.Length)
             {
                 subsetFont.GlyphLeftSideBearings[newIndex] = originalFont.GlyphLeftSideBearings[oldIndex];
             }
@@ -133,10 +134,10 @@ public class FontSubsetter
             subsetFont.Glyphs = new GlyphData[subsetFont.GlyphCount];
             foreach (var kvp in glyphMapping)
             {
-                ushort oldIndex = kvp.Key;
+                uint oldIndex = kvp.Key;
                 ushort newIndex = kvp.Value;
 
-                if (oldIndex < originalFont.Glyphs.Length)
+                if (oldIndex < (uint)originalFont.Glyphs.Length)
                 {
                     subsetFont.Glyphs[newIndex] = originalFont.Glyphs[oldIndex];
                 }
